refactoring captureparam
- move additional methods for feedback version to subclass
