+++ e7awgsw/simplemulti/capturectrl.py
@@ -6,7 +6,7 @@ from e7awgsw.hwparam import NUM_SAMPLES_IN_ADC_WORD, CAPTURED_SAMPLE_SIZE, CLASS
 from e7awgsw.simplemulti.memorymap import CaptureMasterCtrlRegs, CaptureCtrlRegs, CaptureParamRegs
 from e7awgsw.simplemulti.udpaccess import CaptureRegAccess, WaveRamAccess
 from e7awgsw.simplemulti.hwdefs import DspUnit, CaptureUnit, CaptureModule, AWG, CaptureErr
-from e7awgsw.simplemulti.captureparam import CaptureParam
+from e7awgsw.captureparam import CaptureParam
 from e7awgsw.exception import CaptureUnitTimeoutError
 from e7awgsw.logger import get_file_logger, get_null_logger, log_error, log_warning
 from e7awgsw.lock import ReentrantFileLock
